[devscout] moves dkimQuotedPrintableDecode to SignatureRecord
Breaking api change

dkimQuotedPrintableDecode was moved from SignatureRecordImpl to SignatureRecord since it is relevant to all SignatureRecords not only to our implementation.
While technically a public API I don't expect many users of the lib to call this directly.
Updating the static import should be acceptable for the few who do.

This led to a reorganization and a modernization of the tests.

The tests for the decoding are now on SignatureRecordTest while the tests of SignatureRecordImpl specificities were moved to SignatureRecordImplTest.
