+++ main/src/test/java/org/apache/james/jdkim/SignatureRecordImplTest.java
@@ -21,87 +21,92 @@ package org.apache.james.jdkim;
 
 import org.apache.james.jdkim.api.SignatureRecord;
 import org.apache.james.jdkim.tagvalue.SignatureRecordImpl;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
+
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
 import org.junit.Test;
 
 public class SignatureRecordImplTest {
-
     @Test
-    public void testQPDecode() {
-        assertEquals("", SignatureRecordImpl.dkimQuotedPrintableDecode(""));
-        assertEquals("@", SignatureRecordImpl.dkimQuotedPrintableDecode("=40"));
-        assertEquals("\r\n", SignatureRecordImpl
-                .dkimQuotedPrintableDecode("=0D=0A"));
-        assertEquals("\0CIAO\0", SignatureRecordImpl
-                .dkimQuotedPrintableDecode("=00CIAO=00"));
-        assertEquals("thisisatest", SignatureRecordImpl
-                .dkimQuotedPrintableDecode("this\r\n\tis\r\n a\r\n  \t test"));
+    public void testBasic() {
+        new SignatureRecordImpl(
+                "v=1; a=rsa-sha256; c=relaxed/relaxed;\r\n"
+                        + "        d=gmail.com; s=beta;\r\n"
+                        + "        h=domainkey-signature:received:received:message-id:date:from:to:subject:mime-version:content-type;\r\n"
+                        + "        bh=9sd6eO/xnGLInYGPFN86r9q27iClGpwfkl4PBc5XEuQ=;\r\n"
+                        + "        b=tGQtBQg1sO+JKopOylApWLngylEqeMcXwCEUQN+S2PSpi9c1G9Nm5df9pMShus3iFaQb0PPvTfpw++cAC8/N0p3Gi/lVLc+Yh7xWEIPZ3Nxd3xqTQy7grIkBpV0q6559dEhhfFoEyLS0OK/IrqFIUVDRIMnsMjimXV7u+Sgoi7Q=")
+                .validate();
+
     }
 
     @Test
-    public void testQPWhiteSpaces() {
-        assertEquals("thisisatest", SignatureRecordImpl
-                .dkimQuotedPrintableDecode("this is a test"));
-        assertEquals("thisisatest", SignatureRecordImpl
-                .dkimQuotedPrintableDecode("this\r\n is a test"));
+    public void testWrongOrMissingVersion() {
+        assertThatThrownBy(() ->
+                new SignatureRecordImpl(
+                        "a=rsa-sha1; c=relaxed/relaxed;\r\n"
+                                + "        d=gmail.com; s=beta;\r\n"
+                                + "        h=domainkey-signature:received:received:message-id:date:from:to:subject:mime-version:content-type;\r\n"
+                                + "        b=Kw/TqnjB4L5ZC7DX1ibiNkuIw630uHZvzuozn/e6yTm3U8ObWEz/rJK5GO8RSrF56JrCA/xo8W2CGmyNmpQYbEpLl5P9/NcJSYUmln/O6GSa4Usyv4FdEU4FVjkyW0ToGFHNkw9Mm0urveA4Lcfk9gClJczXnvGBdiv/bkVBEJk=")
+                        .validate()
+        ).isInstanceOf(IllegalStateException.class)
+                .hasMessageContaining("mandatory")
+                .hasMessageContaining("v");
+        assertThatThrownBy(() ->
+                new SignatureRecordImpl(
+                        "v=2; a=rsa-sha256; c=relaxed/relaxed;\r\n"
+                                + "        d=gmail.com; s=beta;\r\n"
+                                + "        h=domainkey-signature:received:received:message-id:date:from:to:subject:mime-version:content-type;\r\n"
+                                + "        bh=9sd6eO/xnGLInYGPFN86r9q27iClGpwfkl4PBc5XEuQ=;\r\n"
+                                + "        b=tGQtBQg1sO+JKopOylApWLngylEqeMcXwCEUQN+S2PSpi9c1G9Nm5df9pMShus3iFaQb0PPvTfpw++cAC8/N0p3Gi/lVLc+Yh7xWEIPZ3Nxd3xqTQy7grIkBpV0q6559dEhhfFoEyLS0OK/IrqFIUVDRIMnsMjimXV7u+Sgoi7Q=")
+                        .validate()
+        ).isInstanceOf(IllegalStateException.class)
+                .hasMessageContaining("Invalid")
+                .hasMessageContaining("version")
+                .hasMessageContaining("2");
     }
 
     @Test
-    public void testQPInvalid() {
-        try {
-            SignatureRecordImpl.dkimQuotedPrintableDecode("=");
-            fail("invalid sequence parsed.");
-        } catch (IllegalArgumentException e) {
-        }
-        try {
-            SignatureRecordImpl.dkimQuotedPrintableDecode("==");
-            fail("invalid sequence parsed.");
-        } catch (IllegalArgumentException e) {
-        }
-        try {
-            SignatureRecordImpl.dkimQuotedPrintableDecode("=2 3");
-            fail("invalid sequence parsed.");
-        } catch (IllegalArgumentException e) {
-        }
-        try {
-            SignatureRecordImpl.dkimQuotedPrintableDecode("=3");
-            fail("invalid sequence parsed.");
-        } catch (IllegalArgumentException e) {
-        }
-        try {
-            SignatureRecordImpl.dkimQuotedPrintableDecode("=3a");
-            fail("invalid sequence parsed.");
-        } catch (IllegalArgumentException e) {
-        }
-        try {
-            SignatureRecordImpl.dkimQuotedPrintableDecode("==20");
-            fail("invalid sequence parsed.");
-        } catch (IllegalArgumentException e) {
-        }
-        try {
-            SignatureRecordImpl.dkimQuotedPrintableDecode("=20=");
-            fail("invalid sequence parsed.");
-        } catch (IllegalArgumentException e) {
-        }
-        try {
-            SignatureRecordImpl.dkimQuotedPrintableDecode("=3x");
-            fail("invalid sequence parsed.");
-        } catch (IllegalArgumentException e) {
-        }
-        try {
-            SignatureRecordImpl.dkimQuotedPrintableDecode("this\r\nis a test");
-            fail("invalid sequence parsed.");
-        } catch (IllegalArgumentException e) {
-        }
+    public void testMissingRequired() {
+        assertThatThrownBy(() ->
+            new SignatureRecordImpl(
+                    "v=1; a=rsa-sha256; c=relaxed/relaxed;\r\n"
+                            + "        d=gmail.com; s=beta;\r\n"
+                            + "        h=domainkey-signature:received:received:message-id:date:from:to:subject:mime-version:content-type;\r\n"
+                            + "        b=tGQtBQg1sO+JKopOylApWLngylEqeMcXwCEUQN+S2PSpi9c1G9Nm5df9pMShus3iFaQb0PPvTfpw++cAC8/N0p3Gi/lVLc+Yh7xWEIPZ3Nxd3xqTQy7grIkBpV0q6559dEhhfFoEyLS0OK/IrqFIUVDRIMnsMjimXV7u+Sgoi7Q=")
+                    .validate()
+        ).isInstanceOf(IllegalStateException.class)
+                .hasMessageContaining("mandatory")
+                .hasMessageContaining("bh");
     }
 
-    // TODO check bytes > 128
-    /*
-     * public void test8bit() {
-     * assertEquals("PROVA\144CIAO\144",Main.dkimQuotedPrintableDecode("PROVA=90CIAO=90")); }
-     */
+    @Test
+    public void testDomainMismatch() {
+        assertThatThrownBy(() ->
+                new SignatureRecordImpl(
+                        "v=1; a=rsa-sha256; c=relaxed/relaxed;\r\n"
+                                + "        d=gmail.com; s=beta; i=@agmail.com;\r\n"
+                                + "        h=domainkey-signature:received:received:message-id:date:from:to:subject:mime-version:content-type;\r\n"
+                                + "        bh=9sd6eO/xnGLInYGPFN86r9q27iClGpwfkl4PBc5XEuQ=;\r\n"
+                                + "        b=tGQtBQg1sO+JKopOylApWLngylEqeMcXwCEUQN+S2PSpi9c1G9Nm5df9pMShus3iFaQb0PPvTfpw++cAC8/N0p3Gi/lVLc+Yh7xWEIPZ3Nxd3xqTQy7grIkBpV0q6559dEhhfFoEyLS0OK/IrqFIUVDRIMnsMjimXV7u+Sgoi7Q=")
+                        .validate()
+        ).isInstanceOf(IllegalStateException.class)
+                .hasMessageContaining("domain mismatch");
+    }
+
+    @Test
+    public void testMissingFrom() {
+        assertThatThrownBy(() ->
+                new SignatureRecordImpl(
+                        "v=1; a=rsa-sha256; c=relaxed/relaxed;\r\n"
+                                + "        d=gmail.com; s=beta; i=@subdomain.gmail.com;\r\n"
+                                + "        h=domainkey-signature:received:received:message-id:date:fram:to:subject:mime-version:content-type;\r\n"
+                                + "        bh=9sd6eO/xnGLInYGPFN86r9q27iClGpwfkl4PBc5XEuQ=;\r\n"
+                                + "        b=tGQtBQg1sO+JKopOylApWLngylEqeMcXwCEUQN+S2PSpi9c1G9Nm5df9pMShus3iFaQb0PPvTfpw++cAC8/N0p3Gi/lVLc+Yh7xWEIPZ3Nxd3xqTQy7grIkBpV0q6559dEhhfFoEyLS0OK/IrqFIUVDRIMnsMjimXV7u+Sgoi7Q=")
+                        .validate()
+        ).isInstanceOf(IllegalStateException.class)
+                .hasMessage("From field not signed");
+    }
+
 
     /*
      * when we moved from Sun's Base64 to CommonsCodec the decoding changed
@@ -116,37 +121,21 @@ public class SignatureRecordImplTest {
 
     @Test
     public void testWrongHashSyntaxes() {
-        SignatureRecord sr = new SignatureRecordImpl("v=1; a=nothyphenedword;");
-        try {
-            sr.getHashAlgo();
-            fail("expected failure");
-        } catch (Exception e) {
-            assertTrue(e.getMessage().toLowerCase().contains("hash"));
-        }
-        try {
-            sr.getHashMethod();
-            fail("expected failure");
-        } catch (Exception e) {
-            assertTrue(e.getMessage().toLowerCase().contains("hash"));
-        }
-        try {
-            sr.getHashAlgo();
-            fail("expected failure");
-        } catch (Exception e) {
-            assertTrue(e.getMessage().toLowerCase().contains("hash"));
-        }
+        SignatureRecord sr = new SignatureRecordImpl("v=1; a=nothyphenedword; bh=abcdef; b=1235345987; h=from:to; s=select; d=example.com");
+        assertThatThrownBy(sr::getHashAlgo)
+                .hasMessageContaining("hash");
+
+        assertThatThrownBy(sr::getHashMethod)
+                .hasMessageContaining("hash");
     }
 
     @Test
     public void testExpired() {
-        SignatureRecord sr = new SignatureRecordImpl(
-                "v=1; c=simple; h=from:to; s=select; d=example.com; a=rsa-sha1; x=0; bh=abcdef; b=1235345987;");
-        try {
-            sr.validate();
-            fail("expected failure");
-        } catch (Exception e) {
-            assertTrue(e.getMessage().contains("expired"));
-        }
+        assertThatThrownBy(() ->
+                new SignatureRecordImpl(
+                        "v=1; c=simple; h=from:to; s=select; d=example.com; a=rsa-sha1; x=0; bh=abcdef; b=1235345987;"
+                ).validate()
+        ).hasMessageContaining("expired");
     }
 
 }
